+++ tests/test_app/tests/utils.py
@@ -1,4 +1,4 @@
-# ruff: noqa: N802, RUF012
+# ruff: noqa: N802, RUF012, T201
 import asyncio
 import os
 import sys
@@ -117,16 +117,8 @@ class PlaywrightTestCase(ChannelsLiveServerTestCase):
         cls.browser = cls.playwright.chromium.launch(headless=bool(headless))
         cls.page = cls.browser.new_page()
         cls.page.set_default_timeout(10000)
-        cls.page.on("console", cls.playwright_logging)
-
-    @staticmethod
-    def playwright_logging(msg):
-        if msg.type == "error":
-            _logger.error(msg.text)
-        elif msg.type == "warning":
-            _logger.warning(msg.text)
-        elif msg.type == "info":
-            _logger.info(msg.text)
+        cls.page.on("console", lambda msg: print(f"CLIENT {msg.type.upper()}: {msg.text}"))
+        cls.page.on("pageerror", lambda err: print(f"CLIENT EXCEPTION: {err.name}: {err.message}\n{err.stack}"))
 
     @classmethod
     def shutdown_playwright_client(cls):